middle/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
azure-ai-projects>=2.0.0
azure-identity>=1.20.0
